agentic-scholarships/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
Werkzeug==3.0.4
SQLAlchemy==2.0.35
alembic==1.13.2
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
PyJWT==2.9.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
requests==2.32.3
langchain-core==0.2.39
openai==1.51.0
